feat: add confirmation before deleting an event

+++ src/components/ui/DeleteFab.js
@@ -3,6 +3,7 @@ import 'react-tiny-fab/dist/styles.css';
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { Fab } from 'react-tiny-fab';
+import Swal from 'sweetalert2';
 
 import { deleteEvent } from '../actions/event';
 
@@ -10,7 +11,17 @@ export const DeleteFab = () => {
   const dispatch = useDispatch();
 
   const handleClick = () => {
-    dispatch(deleteEvent());
+    Swal.fire({
+      title: 'Are you sure you want to delete the event?',
+      showCancelButton: true,
+      confirmButtonText: 'Delete',
+      icon: 'question',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteEvent());
+        Swal.fire('Deleted!', '', 'success');
+      }
+    });
   };
 
   return (
